Use our Fog fork to handle Keystone v2 and v3 together.
If OpenStack exposes multiple versions of Keystone, the client
needs to decide whether to prepend v2.0 or v3 inside the API URL
path. This is because the endpoint in Keystone's DB needs to have
the version stripped out, otherwise the service catalog would
force the client to use either v2 or v3.

Leaving the decision in the hands of the client is the best way
around this.