Reap stuck IPs back to the pool.
There are three types of stuck IP:

1) Stuck in a project.

A user assigned this to their project but it isn't
currently allocated to an instance and isn't of any
use to anybody. Users may want to hoard specific IPs
but this isn't a practice that should be encouraged.

2) Stuck in an inactive instance

This is held hostage by an instance that isn't active
and if the user only has one instance then the IP is
free so it's costing them nothing to hang onto it.

3) Stuck on a router

Public gateways on routers get public IPs. If a user
is inactive and not paying then we need to reclaim
these IPs.

* Note on changes in config/initializers/openstack.rb

Without these, the instance_resize_test stubs don't unstub.

This means we get mocks being returned from other calls to
the Fog library. It's shit to remove the memoization but
there isn't significant performance benefit from caching
the connection and there are bugs relating to automatic
reconnection in the background.