Prevent users being spammed with usage alert emails.
Trial users have very small quotas and are almost instantly
spammed with usage alert emails. Make it so the email alerts
the user and then store a timestamp in the database of the
current time. This timestamp is then checked whenever we're
sending out new alerts and a new alert is only sent if a
certain amount of time has elapsed.

Once every 90 days is probably enough.