Changes to the format of usage data. (#356)
The format of the data document we use to describe data usage
for an organization is now substantially different.

The reason for this is to reduce the amount of code in Harbour
that has to adjust the data for the user's benefit - we want the
cleanest format possible internally before we expose the data to
the user. This will hopefully keep format changes to a minimum
going forward and allow us to roll out small changes with API
microversioning rather than big-bang style breaking changes.

Highlights:

* Volume usage is now calculated in line with instance usage. This
  means significantly less code as the volume states reflect their
  combined usage totals rather than some complex method that tries
  to wrangle everything together.
* The data is now stored completely in JSON on Ceph, rather than
  in a marshal-dump form containing real Ruby objects. This will
  take up a lot less space and be less error prone in future.
* The general format of the data is cleaner and tidier with more
  consistent features across different resource types. This
  includes a consistent format for usage and costing per resource.

The Billing::Usage objects on production will need rebuilding
manually, and the volume states will need to be reindexed. I will
be performing these changes during a period of scheduled downtime.