Make OpenStack Connections Switchable (#366)
As part of federation we will need to interact with multiple cloud providers.
For now this will be via the Sentinel daemon which requires a very different
client configuration, X.509 vs username and password.  This pushes configuration
all the way down to the config file, and allows ERB in the configuration for
access to environmental data.