Disabling a user stops their access to all projects.
This is bad news if we want them to be able to belong to
multiple organizations.

Instead of disabling them in Keystone, if we determine that
they should not have OpenStack access then we just remove
all their project memberships.

If we want to limit access on an organizational level then we
can still enable/disable projects themselves.